refactor: change Header component to Server component and rename Nav component to HeadActions
